restructure in Package, add tests, change output to directly retrieve a Parameter instance when adding a new Parameter
