+++ .gitignore
@@ -4,7 +4,7 @@ out/
 !**/src/test/**/out/
 *.iml
 /.idea/
-/test/
+/src/test/
 
 ### Eclipse ###
 .apt_generated
